Fixed most dimension issues, notes are now easily visible
Changed dimension of iframe to make all the data in the page visible. Also allowed for scrolling, that can be a temporary fix. Must be noted that the calendar itself functions correctly but since there is still one other issue with the calendar page itself it looks like it is not.